@@ -2,18 +2,17 @@
   "name": "advent-of-code-2022",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
   "scripts": {
     "test": "jest",
-    "new": "node new.js"
+    "new": "ts-node new.ts"
   },
   "license": "ISC",
   "devDependencies": {
-    "@types/jest": "29.2.2",
+    "@types/jest": "29.2.3",
     "@types/node": "18.11.9",
     "jest": "29.3.1",
     "ts-jest": "29.0.3",
     "ts-node": "10.9.1",
-    "typescript": "4.8.4"
+    "typescript": "4.9.3"
   }
 }
